backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# API clients for data ingestion
requests==2.31.0
SPARQLWrapper==2.0.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0
